requirements.txt
langchain
fastapi
uvicorn
pydantic
langchain-google-genai
python-dotenv
openai-whisper
youtube-transcript-api
langchain_community
sentence_transformers
langchain_huggingface
faiss-cpu
pinecone
langchain_pinecone
pinecone_init
